pyproject.toml
[project]
name = "m-rag"
version = "0.1.0"
description = "Multimodal RAG Framework — PDF + Image ingestion with LangGraph"
requires-python = ">=3.12"
dependencies = [
    # LLM / LangChain
    "langchain",
    "langchain-groq",
    "langchain-community",
    "langgraph",
    # Vector store
    "qdrant-client",
    "langchain-qdrant",
    # Postgres / ORM
    "sqlmodel",
    "asyncpg",
    "psycopg2-binary",
    # Embeddings
    "sentence-transformers",
    # Document parsing
    # "unstructured[pdf,image]",
    "pymupdf",
    "pillow",
    # Observability
    "langfuse",
    # CLI
    "typer[standard]",
    "rich",
    # Async / utilities
    "httpx",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "tenacity",
    # Reranker
    "sentence-transformers @ git+https://github.com/UKPLab/sentence-transformers",]

[project.scripts]
m-rag = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "core", "ingestion", "rag", "scripts", "utils"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }